users_have_apps_table_latitude_longitude_fixed

+++ database/migrations/2019_12_30_203210_create_users_have_apps_table.php
@@ -21,8 +21,8 @@ class CreateUsersHaveAppsTable extends Migration
             $table->foreign('app_id')->references('id')->on('apps')->onDelete('cascade');
             $table->dateTime('date'); 
             $table->string('event');           
-            $table->integer('latitude');
-            $table->integer('longitude');
+            $table->double('latitude', 8, 6);
+            $table->double('longitude', 8, 6);
             $table->timestamps();
         });
     }
